move promise to utils/ dir
