+++ test/HighTable.test.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { DataFrame, sortableDataFrame } from '../src/dataframe.js'
 import HighTable from '../src/HighTable.js'
-import { wrapPromise } from '../src/promise.js'
+import { wrapPromise } from '../src/utils/promise.js'
 import { render } from './userEvent.js'
 
 const data: DataFrame = {
