+++ test/rowCache.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, vi } from 'vitest'
-import { wrapPromise } from '../src/promise.js'
 import { AsyncRow, awaitRows } from '../src/row.js'
 import { rowCache } from '../src/rowCache.js'
+import { wrapPromise } from '../src/utils/promise.js'
 
 // Mock DataFrame
 function makeDf() {
